IggyTools/requirements.txt
MySQL-python>=1.2.5
PyYAML>=3.11
SQLAlchemy>=0.9.7
argparse>=1.2.1
